Buffer the bytesused variable from struct v4l2_buffer
Starting with kernel versions 3.16, (DE)Queing of buffers has been fixed
after it was leaking data for ages. in the struct v4l2_buffer is the
bytesused element which indicates the size of the buffer. This however
gets cleared whenever the buffer gets requeued and is thus no longer
valid.

This patch copies the bytesused variable so it is available until the
next frame captured again.

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>