Fix output_viewer plugin (multiple compile issues)
The output_viewer plugin wasn't compiled in a while it seems because it
did not function. This patch fixes the compiler and linker issues
(ldflags needs to be on the end for some reason).

There is still a memory leak, as the viewer starts, displays the feed
for a few seconds, and then segfaults. TODO ;)

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>